Refactor PyWindow to be moved to Expr submodule
