+++ src/expr.rs
@@ -662,5 +662,8 @@ pub(crate) fn init_module(m: &PyModule) -> PyResult<()> {
     m.add_class::<drop_table::PyDropTable>()?;
     m.add_class::<repartition::PyPartitioning>()?;
     m.add_class::<repartition::PyRepartition>()?;
+    m.add_class::<window::PyWindow>()?;
+    m.add_class::<window::PyWindowFrame>()?;
+    m.add_class::<window::PyWindowFrameBound>()?;
     Ok(())
 }
